+++ netdef/Engines/webadmin/SecurityBuiltinControllersView.py
@@ -113,3 +113,5 @@ class SecurityBuiltinControllersView(MyBaseView):
             conf_ok=conf_ok,
             form=form
         )
+    def is_accessible(self):
+        return super().is_accessible() and self.has_role("admin")